remove unused hc zome call. create shared func for kyc
